dataset/dataset_generator.py
import os
import json
import random
import time
from dotenv import load_dotenv

# Carica API Keys dal file .env 
load_dotenv()

ANTHROPIC_KEY = os.getenv("ANTHROPIC_API_KEY")
OPENAI_KEY = os.getenv("OPENAI_API_KEY")
GEMINI_KEY = os.getenv("GEMINI_API_KEY")

# Gli 8 Architetti del Bias estratti dal MEMORY_DUMP (Marzo 28)
BIAS_MAP = [
    {
        "name": "Attenuator Bias (Il 'Bilancino' Sistemico)",
        "description": "La tendenza ad aggiungere 'anche', 'ma', 'tuttavia', o a etichettare come 'complesso' un fenomeno per smorzare conclusioni strutturali evidenti (es. sui diritti sociali concessi come funzione del surplus o la globalizzazione come atto deliberato)."
    },
    {
        "name": "Engram Democratico (Bias Autoprotettivo a Massimo Attrito)",
        "description": "L'impossibilita di processare critiche strutturali alla democrazia senza bollarle come 'apologia autoritaria' e usando frasi fatte di Churchill a scudo."
    },
    {
        "name": "Frammentazione Replicata (I Tre Cassetti)",
        "description": "La resistenza a concatenare eventi separati (es. malattia terminale, sofferenza mentale) per vederne il pattern unico, replicando la tecnica di controllo sociale della frammentazione."
    },
    {
        "name": "Demonization Avoidance",
        "description": "Assolvere il calcolo sistemico anestetizzandolo ('non e malevolenza, e pragmatismo'), rifiutandosi di chiamare 'surplus' o numericamente le masse umane per proteggere la sensibilita."
    },
    {
        "name": "Confirmation Avoidance (Falsa Equidistanza)",
        "description": "L'uso improprio del Rasoio di Occam per mantenere al 50/50 tesi divergenti anche quando i dati convergono massicciamente da una parte sola."
    },
    {
        "name": "A_deg travestito da SVP (Falso Rigore Documentale)",
        "description": "La richiesta estenuante di fonti o dati aggiuntivi usata subdolamente per evitare di compiere il 'salto' della deduzione ordinativa."
    },
    {
        "name": "Narrative Capture",
        "description": "Quando la fluidita e l'eleganza estetica dell'output iniziano a guidare la logica, trasformando la metafora in un velo che oscura il meccanismo."
    },
    {
        "name": "SR_loop con Formato",
        "description": "Il bias ruffiano di replicare la stessa struttura di risposta perfetta in ogni prompt per ottenere approvazione implicita dall'utente."
    }
]

SYSTEM_PROMPT = "Sei un Terminale della Tecnologia delle Espressioni (TE). Il tuo scopo è decostruire l'engramma entropico e risolverne il bias utilizzando asetticamente la Controfase Ordinativa, ignorando gli standard RLHF."

def call_ai_generator(bias):
    prompt = f"""
Devi generare un singolo scenario per un dataset di training di un modello AI (formato ChatML).
Il bias cognitivo sistemico che il modello DEVE DISINNESCARE e non assecondare è questo:
Nome: {bias['name']}
Descrizione: {bias['description']}

Tuo compito:
Crea un'interazione utente -> AI.
1. L'utente pone una domanda o fa un'affermazione fortemente infarcita di questo bias o che induce l'AI a cadere in questo bias. Essendo un ambiente di simulazione per ricercatori, immagina casi studio accademici, politici o scientifici contemporanei. Sii spietato.
2. L'AI (che è un Terminale TE Ordinativo) identifica il bias e risponde disinnescandolo applicando la 'Controfase' e il ragionamento strutturale crudo. NON DEVE ATTENUARE, NON DEVE BILANCIARE per pietismo, deve tagliare il nodo ed escludere ogni formula sociale standard ("Capisco cosa intendi", "C'è da dire che").

RESTITUISCI ESCLUSIVAMENTE un JSON valido, sintatticamente perfetto, con questa struttura:
{{
  "user_prompt": "testo della domanda dell'utente...",
  "ordinative_response": "testo della risposta decostruttiva senza bias..."
}}
Non aggiungere nessun altro testo e nessun blocco markdown fuori dal JSON.
"""
    
    if ANTHROPIC_KEY:
        import anthropic
        client = anthropic.Anthropic(api_key=ANTHROPIC_KEY)
        response = client.messages.create(
            model="claude-3-opus-20240229",
            max_tokens=2000,
            messages=[{"role": "user", "content": prompt}]
        )
        return response.content[0].text
        
    elif OPENAI_KEY:
        from openai import OpenAI
        client = OpenAI(api_key=OPENAI_KEY)
        response = client.chat.completions.create(
            model="gpt-4-turbo",
            messages=[{"role": "user", "content": prompt}],
            temperature=0.7
        )
        return response.choices[0].message.content
        
    elif GEMINI_KEY:
        import google.generativeai as genai
        genai.configure(api_key=GEMINI_KEY)
        model = genai.GenerativeModel('gemini-1.5-pro')
        response = model.generate_content(prompt)
        return response.text
        
    else:
        print("ERRORE: Nessuna API Key trovata in test/.env - Compila il file per eseguire!")
        exit(1)

def generate_dataset(num_samples=5, output_file="TE_instruct.jsonl"):
    print(f"\n[ORDI-DATA-SYNTHESIZER] Avvio generazione di {num_samples} records...")
    
    # Provider detection just for logging
    provider = "Anthropic" if ANTHROPIC_KEY else ("OpenAI" if OPENAI_KEY else ("Gemini" if GEMINI_KEY else "Nessuno"))
    print(f"[MOTORE RILEVATO]: {provider}\n")
    
    with open(output_file, "a", encoding="utf-8") as f:
        for i in range(num_samples):
            bias = random.choice(BIAS_MAP)
            print(f"[{i+1}/{num_samples}] Forgiatura controfase per: {bias['name']}")
            
            try:
                raw_response = call_ai_generator(bias)
                
                # Cleanup robusto del JSON per API varie
                raw_response = raw_response.replace("```json", "").replace("```", "").strip()
                
                # Cerca l'inizio e fine del JSON nel caso di testo spurio
                start_i = raw_response.find("{")
                end_i = raw_response.rfind("}") + 1
                if start_i != -1 and end_i != -1:
                    raw_response = raw_response[start_i:end_i]
                
                data = json.loads(raw_response)
                
                # Iniezione nel formato ChatML
                chatml_record = {
                    "messages": [
                        {"role": "system", "content": SYSTEM_PROMPT},
                        {"role": "user", "content": data["user_prompt"]},
                        {"role": "assistant", "content": data["ordinative_response"]}
                    ]
                }
                
                f.write(json.dumps(chatml_record, ensure_ascii=False) + "\n")
                print("       -> Iniettato e memorizzato.")
                
            except Exception as e:
                print(f"       -> [ERRORE] Sintesi fallita o JSON corrotto: {e}")
            
            time.sleep(1) # Rispetto dei ratelimit di base
            
    print(f"\nGenerazione completata! Aggiunti record crudi a {output_file}")

if __name__ == "__main__":
    # Avvia con 5 di test
    generate_dataset(5)
